requirements.txt
Django<4.0
djangorestframework==3.12
djangorestframework-xml
django-oauth-toolkit
django-cors-headers
Pillow
mysqlclient
django-pipeline
sorl-thumbnail
tld
lxml
qiniu
elasticsearch==6.3.0
redis<4.0.0
hashids
xmltodict
python-docx
python-dateutil
xlrd
xpinyin
channels<4.0.0
channels_redis
django-celery-results
django-celery-beat
django-redis
celery<5.0.0
flower
redis-py-cluster<3.0.0
requests
numpy
openpyxl

